# server/pkg/api/rest/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collecting and Aggregating server module",
        "title": "CM-Honeybee Server REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/honeybee",
    "paths": {
...
        "/source_group/{sgId}/connection_info/{connId}/infra": {
            "get": {
                "description": "Get the infra information of the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Get] Get source info"
                ],
                "summary": "Get Infra Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connection info.",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the infra.",
                        "schema": {
                            "$ref": "#/definitions/infra.Infra"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the infra.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source_group/{sgId}/connection_info/{connId}/software": {
            "get": {
                "description": "Get the software information of the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Get] Get source info"
                ],
                "summary": "Get Software Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connection info.",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the software.",
                        "schema": {
                            "$ref": "#/definitions/software.Software"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the software.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/source_group/{sgId}/infra": {
            "get": {
                "description": "Get the infra information for all connections in the source group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Get] Get source info"
                ],
                "summary": "Get Infra Information Source Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the infra.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.InfraInfoList"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the infra.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source_group/{sgId}/software": {
            "get": {
                "description": "Get the software information for all connections in the source group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Get] Get source info"
                ],
                "summary": "Get Software Information Source Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the software.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SoftwareInfoList"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the infra.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.InfraInfoList": {
            "type": "object",
            "required": [
                "servers"
            ],
            "properties": {
                "servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.Infra"
                    }
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SoftwareInfoList": {
            "type": "object",
            "required": [
                "servers"
            ],
            "properties": {
                "servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/software.Software"
                    }
                }
            }
        },
...
        "infra.CPU": {
            "type": "object",
            "required": [
                "cores",
                "cpus",
                "threads"
            ],
            "properties": {
                "cache": {
                    "description": "KB",
                    "type": "integer"
                },
                "cores": {
                    "description": "ea",
                    "type": "integer"
                },
                "cpus": {
                    "description": "ea",
                    "type": "integer"
                },
                "max_speed": {
                    "description": "MHz",
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "threads": {
                    "description": "ea",
                    "type": "integer"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "infra.Compute": {
            "type": "object",
            "required": [
                "compute_resource",
                "os"
            ],
            "properties": {
                "compute_resource": {
                    "$ref": "#/definitions/infra.ComputeResource"
                },
                "connection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.Connection"
                    }
                },
                "os": {
                    "$ref": "#/definitions/infra.System"
                }
            }
        },
        "infra.ComputeResource": {
            "type": "object",
            "required": [
                "cpu",
                "memory"
            ],
            "properties": {
                "cpu": {
                    "$ref": "#/definitions/infra.CPU"
                },
                "data_disk": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.Disk"
                    }
                },
                "memory": {
                    "$ref": "#/definitions/infra.Memory"
                },
                "root_disk": {
                    "$ref": "#/definitions/infra.Disk"
                }
            }
        },
        "infra.Connection": {
            "type": "object",
            "properties": {
                "keypair": {
                    "$ref": "#/definitions/infra.Keypair"
                }
            }
        },
        "infra.DRM": {
            "type": "object",
            "properties": {
                "driver_date": {
                    "type": "string"
                },
                "driver_description": {
                    "type": "string"
                },
                "driver_name": {
                    "type": "string"
                },
                "driver_version": {
                    "type": "string"
                }
            }
        },
        "infra.Disk": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "size": {
                    "description": "GB",
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "infra.GPU": {
            "type": "object",
            "properties": {
                "drm": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.DRM"
                    }
                },
                "nvidia": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.NVIDIA"
                    }
                }
            }
        },
        "infra.Infra": {
            "type": "object",
            "properties": {
                "compute": {
                    "$ref": "#/definitions/infra.Compute"
                },
                "gpu": {
                    "$ref": "#/definitions/infra.GPU"
                },
                "network": {
                    "$ref": "#/definitions/network.Network"
                },
                "storage": {
                    "$ref": "#/definitions/infra.Storage"
                }
            }
        },
        "infra.Kernel": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "infra.Keypair": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                }
            }
        },
        "infra.Memory": {
            "type": "object",
            "required": [
                "size"
            ],
            "properties": {
                "size": {
                    "description": "MB",
                    "type": "integer"
                },
                "speed": {
                    "description": "MHz",
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "infra.MountPoint": {
            "type": "object",
            "properties": {
                "mounted_information": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/infra.MountedInformation"
                    }
                }
            }
        },
...
        "infra.NVIDIA": {
            "type": "object",
            "properties": {
                "device_attribute": {
                    "$ref": "#/definitions/infra.NVIDIADeviceAttribute"
                },
                "performance": {
                    "$ref": "#/definitions/infra.NVIDIAPerformance"
                }
            }
        },
        "infra.NVIDIADeviceAttribute": {
            "type": "object",
            "properties": {
                "cuda_version": {
                    "type": "string"
                },
                "driver_version": {
                    "type": "string"
                },
                "gpu_uuid": {
                    "type": "string"
                },
                "product_architecture": {
                    "type": "string"
                },
                "product_brand": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
        "infra.NVIDIAPerformance": {
            "type": "object",
            "properties": {
                "bar1_memory_total": {
                    "description": "mb",
                    "type": "integer"
                },
                "bar1_memory_usage": {
                    "description": "percent",
                    "type": "integer"
                },
                "bar1_memory_used": {
                    "description": "mb",
                    "type": "integer"
                },
                "fb_memory_total": {
                    "description": "mb",
                    "type": "integer"
                },
                "fb_memory_usage": {
                    "description": "percent",
                    "type": "integer"
                },
                "fb_memory_used": {
                    "description": "mb",
                    "type": "integer"
                },
                "gpu_usage": {
                    "description": "percent",
                    "type": "integer"
                }
            }
        },
        "infra.Node": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "hypervisor": {
                    "type": "string"
                },
                "machineid": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "infra.OS": {
            "type": "object",
            "required": [
                "release",
                "vendor"
            ],
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "infra.Storage": {
            "type": "object",
            "properties": {
                "mount_point": {
                    "$ref": "#/definitions/infra.MountPoint"
                }
            }
        },
        "infra.System": {
            "type": "object",
            "required": [
                "os"
            ],
            "properties": {
                "kernel": {
                    "$ref": "#/definitions/infra.Kernel"
                },
                "node": {
                    "$ref": "#/definitions/infra.Node"
                },
                "os": {
                    "$ref": "#/definitions/infra.OS"
                }
            }
        },
...
        "network.CSP": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "nlb": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.NLB"
                    }
                },
                "security_group": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.SecurityGroup"
                    }
                },
                "vpc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.VPC"
                    }
                }
            }
        },
...
        "network.Host": {
            "type": "object",
            "properties": {
                "dns": {
                    "$ref": "#/definitions/network.DNS"
                },
                "firewall_rule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.FirewallRule"
                    }
                },
                "network_interface": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.NIC"
                    }
                },
                "route": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.Route"
                    }
                }
            }
        },
...
        "network.NLB": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "health_checker": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "listener": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "target_group": {
                    "type": "string"
                }
            }
        },
        "network.Network": {
            "type": "object",
            "properties": {
                "csp": {
                    "$ref": "#/definitions/network.CSP"
                },
                "host": {
                    "$ref": "#/definitions/network.Host"
                }
            }
        },
...
        "network.SecurityGroup": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "firewall_rule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.FirewallRule"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "vnet_id": {
                    "type": "string"
                }
            }
        },
...
        "software.DEB": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "conffiles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "depends": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "homepage": {
                    "type": "string"
                },
                "installed_size": {
                    "type": "integer"
                },
                "maintainer": {
                    "type": "string"
                },
                "multi_arch": {
                    "type": "string"
                },
                "package": {
                    "type": "string"
                },
                "pre_depends": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "software.Docker": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Container"
                    }
                }
            }
        },
        "software.Podman": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Container"
                    }
                }
            }
        },
        "software.RPM": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "requires": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "sourceRpm": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "software.Software": {
            "type": "object",
            "properties": {
                "deb": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/software.DEB"
                    }
                },
                "docker": {
                    "$ref": "#/definitions/software.Docker"
                },
                "podman": {
                    "$ref": "#/definitions/software.Podman"
                },
                "rpm": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/software.RPM"
                    }
                }
            }
        },
        "types.Container": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "created": {
                    "type": "integer"
                },
                "hostConfig": {
                    "type": "object",
                    "properties": {
                        "annotations": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        },
                        "networkMode": {
                            "type": "string"
                        }
                    }
                },
                "image": {
                    "type": "string"
                },
                "imageID": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MountPoint"
                    }
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "networkSettings": {
                    "$ref": "#/definitions/types.SummaryNetworkSettings"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Port"
                    }
                },
                "sizeRootFs": {
                    "type": "integer"
                },
                "sizeRw": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}